Fix: bad clear of indents and refactor on leading widgets
