+++ lib/src/editor/widgets/default_leading_components/bullet_point_leading.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+import '../../raw_editor/builders/config/leading_configurations.dart';
+import '../../style_widgets/style_widgets.dart';
+
+Widget bulletPointLeading(LeadingConfigurations config) =>
+    QuillEditorBulletPoint(
+      style: config.style!,
+      width: config.width!,
+      padding: config.padding!,
+    );
