+++ lib/src/editor/widgets/default_leading_components/number_point_leading.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import '../../raw_editor/builders/config/leading_configurations.dart';
+import '../../style_widgets/style_widgets.dart';
+
+Widget numberPointLeading(LeadingConfigurations config) =>
+    QuillEditorNumberPoint(
+      index: config.getIndexNumberByIndent!,
+      indentLevelCounts: config.indentLevelCounts,
+      count: config.count,
+      style: config.style!,
+      attrs: config.attrs,
+      width: config.width!,
+      padding: config.padding!,
+    );
